# gateway/api/v1/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "即时通讯系统后端接口",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/app/mute-list": {
            "post": {
                "tags": [
                    "group 禁言"
                ],
                "summary": "查询群内被禁言成员名单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MOCK",
                        "name": "FZM-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.GetMuteListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.GeneralResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.GetMuteListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/group/app/group-exit": {
            "post": {
                "tags": [
                    "group 群动作"
                ],
                "summary": "退群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MOCK",
                        "name": "FZM-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.GroupExitReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.GeneralResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.GroupExitResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/group/app/group-pub-info": {
            "post": {
                "tags": [
                    "group 群信息"
                ],
                "summary": "查询群公开信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MOCK",
                        "name": "FZM-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.GetGroupPubInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.GeneralResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.GetGroupPubInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/group/app/join-group": {
            "post": {
                "tags": [
                    "group 群动作"
                ],
                "summary": "直接进群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MOCK",
                        "name": "FZM-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.JoinGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.GeneralResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.JoinGroupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.GeneralResp": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "integer"
                },
                "result": {
                    "type": "integer"
                }
            }
        },
...
        "types.GetGroupPubInfoReq": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "群 ID",
                    "type": "integer"
                },
                "idStr": {
                    "description": "如果同时填了 idStr, 则优先选择 idStr",
                    "type": "string"
                }
            }
        },
        "types.GetGroupPubInfoResp": {
            "type": "object",
            "properties": {
                "adminNum": {
                    "description": "群内管理员数量",
                    "type": "integer"
                },
                "avatar": {
                    "description": "头像 url",
                    "type": "string"
                },
                "createTime": {
                    "description": "群创建时间",
                    "type": "integer"
                },
                "friendType": {
                    "description": "加好友限制， 0=群内可加好友，1=群内禁止加好友",
                    "type": "integer"
                },
                "groupType": {
                    "description": "群类型 (0: 普通群, 1: 全员群, 2: 部门群)",
                    "type": "integer"
                },
                "id": {
                    "description": "群 ID",
                    "type": "integer"
                },
                "idStr": {
                    "type": "string"
                },
                "introduce": {
                    "type": "string"
                },
                "joinType": {
                    "description": "加群方式，0=无需审批（默认），1=禁止加群，群主和管理员邀请加群, 2=普通人邀请需要审批,群主和管理员直接加群",
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "markId": {
                    "description": "群显示的 ID",
                    "type": "string"
                },
                "maximum": {
                    "description": "群人数上限",
                    "type": "integer"
                },
                "memberNum": {
                    "description": "群人数",
                    "type": "integer"
                },
                "muteNum": {
                    "description": "群内当前被禁言的人数",
                    "type": "integer"
                },
                "muteType": {
                    "description": "禁言， 0=全员可发言， 1=全员禁言(除群主和管理员)",
                    "type": "integer"
                },
                "name": {
                    "description": "群名称 加密的",
                    "type": "string"
                },
                "owner": {
                    "description": "群主 信息",
                    "$ref": "#/definitions/types.GroupMember"
                },
                "person": {
                    "description": "本人在群内的信息",
                    "$ref": "#/definitions/types.GroupMember"
                },
                "publicName": {
                    "description": "公开的群名称 不加密的",
                    "type": "string"
                },
                "status": {
                    "description": "群状态，0=正常 1=封禁 2=解散",
                    "type": "integer"
                }
            }
        },
        "types.GetMuteListReq": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "群 ID",
                    "type": "integer"
                },
                "idStr": {
                    "description": "如果同时填了 idStr, 则优先选择 idStr",
                    "type": "string"
                }
            }
        },
        "types.GetMuteListResp": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.GroupMember"
                    }
                }
            }
        },
...
        "types.GroupExitReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "idStr": {
                    "description": "如果同时填了 idStr, 则优先选择 idStr",
                    "type": "string"
                }
            }
        },
        "types.GroupExitResp": {
            "type": "object"
        },
...
        "types.GroupMember": {
            "type": "object",
            "properties": {
                "memberId": {
                    "description": "用户 ID",
                    "type": "string"
                },
                "memberMuteTime": {
                    "description": "该用户被禁言结束的时间 9223372036854775807=永久禁言",
                    "type": "integer"
                },
                "memberName": {
                    "description": "用户群昵称",
                    "type": "string"
                },
                "memberType": {
                    "description": "用户角色，2=群主，1=管理员，0=群员，10=退群",
                    "type": "integer"
                }
            }
        },
...
        "types.JoinGroupReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "idStr": {
                    "type": "string"
                },
                "inviterId": {
                    "type": "string"
                }
            }
        },
        "types.JoinGroupResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "idStr": {
                    "type": "string"
                }
            }
        },
...
    }
}